pyproject.toml
[project]
name = "fastapi-studyroom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.129.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "sqlalchemy[asyncio]>=2.0.46",
]

[dependency-groups]
dev = [
    "ruff>=0.15.2",
]
